changed routing for video screens + temp fix for video card image on ios
