+++ frontend/package.json
@@ -25,7 +25,10 @@
     "react-native-calendars": "^1.1307.0",
     "react-native-safe-area-context": "4.10.5",
     "react-native-screens": "3.31.1",
-    "expo-image-picker": "~15.0.7"
+    "expo-image-picker": "~15.0.7",
+    "react-native-gesture-handler": "~2.16.1",
+    "expo-image": "~1.13.0",
+    "react-native-webview": "13.8.6"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
